requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyascii==0.3.2
asttokens==3.0.0
autocorrect==2.6.1
beautifulsoup4==4.13.3
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fonttools==4.57.0
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl#sha256=7d6ad14cd5078e53147bfbf70fb9d433c6a3865b695fda2657140bbc59a27e29
graphviz==0.20.3
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.34.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.2
packaging==24.2
panda==0.3.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
preshed==3.0.9
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
